run test_monitor through pytest; fix the test, add flake8 to bench direectory - like PR 891
