+++ baselines/bench/__init__.py
@@ -1,2 +1,3 @@
+# flake8: noqa F403
 from baselines.bench.benchmarks import *
 from baselines.bench.monitor import *
