+++ setup.cfg
@@ -4,4 +4,3 @@ exclude =
     .git,
     __pycache__,
     baselines/ppo1,
-    baselines/bench,
